Save built .whl files to skip 5-hour build on future images

+++ overlay/pwnagotchi/03-build-torch/01-run.sh
@@ -2,4 +2,13 @@
 
 echo "Copying files from $(pwd)"
 
-cp torch-packages ${ROOTFS_DIR}/tmp/torch-packages
+if [ -f torch-packages]; then
+    ls -l torch-packages
+    cp torch-packages ${ROOTFS_DIR}/tmp/torch-packages
+fi
+
+if [ -d INCOMING_files ]; then
+    cp INCOMING_files/* ${ROOTFS_DIR}/tmp
+fi
+
+ls -ltrh ${ROOTFS_DIR}/tmp
